+++ packages/contracts/src/presets.ts
@@ -6,7 +6,7 @@ import {
   meta,
   MonkeyResponseSchema,
   responseWithData,
-} from "./schemas/util";
+} from "./util/api";
 import {
   EditPresetRequestSchema,
   PresetSchema,
