+++ packages/contracts/src/results.ts
@@ -5,7 +5,7 @@ import {
   meta,
   MonkeyResponseSchema,
   responseWithData,
-} from "./schemas/util";
+} from "./util/api";
 import {
   CompletedEventSchema,
   PostResultResponseSchema,
